Partial logic implemented, have to figure this out, ran out of ideas
Pretty much, I coded for just the flagships as a test, it will randomly
select a ship, the problem is that the while loop makes it fly towards
our ship at a crazy speed within half a second.

After we figure that out, it’ll be easy to make it return back into the
fleet position.